pyproject.toml
[project]
name = "ascension-backend"
version = "0.1.0"
description = "Minimal Starlette server for Ascension"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
    "httpx>=0.28.1",
    "google-adk>=1.0.0",
    "python-dotenv>=1.0.1",
    "litellm>=1.0.0",
    "exa-py>=1.0.0",
    "tavily-python>=0.5.0",
    "a2a-sdk>=0.3.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "pytest>=8.0.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\\.(py|pyi)$'
exclude = '''
/( 
  \.venv
  | build
  | dist
  | __pycache__
  | \.egg-info
)/
'''

[tool.isort]
profile = "black"
src_paths = ["backend"]
line_length = 88
py_version = 311
